Fix php build warning #1331

+++ php/src/php/Operation.cpp
@@ -382,7 +382,7 @@ IcePHP::OperationI::getArgInfo(zend_internal_arg_info& arg, const ParamInfoPtr& 
     {
         zend_internal_arg_info ai[] =
         {
-            ZEND_ARG_ARRAY_INFO(static_cast<uint32_t>(pass_by_ref),
+            ZEND_ARG_ARRAY_INFO(pass_by_ref,
                                 static_cast<uint32_t>(0),
                                 static_cast<uint32_t>(allow_null))
         };
@@ -392,7 +392,7 @@ IcePHP::OperationI::getArgInfo(zend_internal_arg_info& arg, const ParamInfoPtr& 
     {
         zend_internal_arg_info ai[] =
         {
-            ZEND_ARG_CALLABLE_INFO(static_cast<uint32_t>(pass_by_ref),
+            ZEND_ARG_CALLABLE_INFO(pass_by_ref,
                                    static_cast<uint32_t>(0),
                                    static_cast<uint32_t>(allow_null))
         };
